@@ -1,3 +1,3 @@
 module github.com/GoogleCloudPlatform/appengine/go11x/tasks/handle_task
 
-go 1.22.0
+go 1.23.0
