@@ -1,3 +1,3 @@
 module github.com/GoogleCloudPlatform/appengine_flexible/helloworld
 
-go 1.22.0
+go 1.23.0
