@@ -27,12 +27,12 @@ import (
 
 // [END gae_go111_import]
 
-// [START main_func]
+// [START gae_go111_main_func]
 
 func main() {
 	http.HandleFunc("/", indexHandler)
 
-	// [START setting_port]
+	// [START gae_go111_setting_port]
 	port := os.Getenv("PORT")
 	if port == "" {
 		port = "8080"
@@ -43,12 +43,12 @@ func main() {
 	if err := http.ListenAndServe(":"+port, nil); err != nil {
 		log.Fatal(err)
 	}
-	// [END setting_port]
+	// [END gae_go111_setting_port]
 }
 
-// [END main_func]
+// [END gae_go111_main_func]
 
-// [START indexHandler]
+// [START gae_go111_indexHandler]
 
 // indexHandler responds to requests with our greeting.
 func indexHandler(w http.ResponseWriter, r *http.Request) {
@@ -59,5 +59,5 @@ func indexHandler(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprint(w, "Hello, World!")
 }
 
-// [END indexHandler]
+// [END gae_go111_indexHandler]
 // [END gae_go111_app]
