@@ -14,7 +14,7 @@
 
 package sample
 
-// [START intro]
+// [START gae_datastore_intro]
 import (
 	"fmt"
 	"net/http"
@@ -57,4 +57,4 @@ func handle(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprintf(w, "Stored and retrieved the Employee named %q", e2.Name)
 }
 
-// [END intro]
+// [END gae_datastore_intro]
